Rename wagmi library files
