+++ src/libs/wagmi/connectors.ts
@@ -12,8 +12,8 @@ import {
   safe,
 } from 'wagmi/connectors';
 import config from 'config';
-import { type SelectableConnectionType } from './web3.types';
-import { selectedConnections } from './web3.constants';
+import { type SelectableConnectionType } from './wagmi.types';
+import { selectedConnections } from './wagmi.constants';
 
 const createPlaceholderConnector = ({
   id,
