+++ src/libs/wagmi/wagmi.types.ts
@@ -4,7 +4,7 @@ import {
   StaticJsonRpcProvider,
   Web3Provider,
 } from '@ethersproject/providers';
-import { SupportedChainId } from 'libs/wagmi/web3.constants';
+import { SupportedChainId } from 'libs/wagmi/wagmi.constants';
 import { Connector } from 'libs/wagmi';
 
 export type ChainIdMapTo<T extends string | JsonRpcProvider> = {
